Метод чтения из файла перенесён в модуль utils
